Now that methods exist, we need to allow stack references within them
The past assumption that we'd only have stack references of "parent" actions is now false, since you can define stack references within methods whose actions get replaced as child actions of the method. To fix these we needed to create a stack history at the end of each action, which required tracking when actions end, and not just when parent actions end.